pyproject.toml
[project]
name = "tanlov"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-unfold>=0.62.0",
    "python-decouple>=3.8",
]
